Create SeekMax.m
This is the meat of the program that goes through digitized data to find the local maxima. In theory this isn't too hard, but with experimental, unsmoothed data finding the maxima is trickier because the data points jump around much more. This code runs in Octave 3.6.4 (and hopefully other distributions) and should run in Matlab as well, possibly with a few small tweaks, though I haven't tried it.